WIP rename a bunch of interfaces and controllers to organize things. VehicleShip becomes VehicleBaseController and there is a shared interface across both BaseController and PiecesController so we can use an interface lookup to get the Pieces or BaseControllers without needing to worry about specific component
