+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 1567431670
+CRC: 3201355153
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
