+++ src/ValheimRAFT/AssetBundle/valheim-vehicles.manifest
@@ -1,15 +1,15 @@
 ManifestFileVersion: 0
-CRC: 4025671589
+CRC: 1096013515
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: 151b6eac3b92da604b9761839f2a34f8
+    Hash: 746cedc2b95c175aeb28ca6b97f0c8ac
   TypeTreeHash:
     serializedVersion: 2
-    Hash: ceb696d42be58906fbd29514cbb24763
+    Hash: cbca681f9021208bbcc6d59371405a02
   IncrementalBuildHash:
     serializedVersion: 2
-    Hash: 151b6eac3b92da604b9761839f2a34f8
+    Hash: 746cedc2b95c175aeb28ca6b97f0c8ac
 HashAppended: 0
 ClassTypes:
 - Class: 1
@@ -59,19 +59,19 @@ ClassTypes:
 - Class: 114
   Script: {fileID: 11500000, guid: d199490a83bb2b844b9695cbf13b01ef, type: 3}
 - Class: 114
+  Script: {fileID: 11500000, guid: 9085046f02f69544eb97fd06b6048fe2, type: 3}
+- Class: 114
+  Script: {fileID: 11500000, guid: 747419968f339df40a827088f25c215d, type: 3}
+- Class: 114
   Script: {fileID: 11500000, guid: 31a19414c41e5ae4aae2af33fee712f6, type: 3}
 - Class: 114
   Script: {fileID: 11500000, guid: f44587a585f05c540ad46e160444cf0c, type: 3}
 - Class: 114
   Script: {fileID: 11500000, guid: 8233d90336aea43098adf6dbabd606a2, type: 3}
 - Class: 114
-  Script: {fileID: 11500000, guid: 9085046f02f69544eb97fd06b6048fe2, type: 3}
-- Class: 114
   Script: {fileID: 11500000, guid: 37593e1934874fe4db91a205982d109e, type: 3}
 - Class: 114
   Script: {fileID: 11500000, guid: 2a4db7a114972834c8e4117be1d82ba3, type: 3}
-- Class: 114
-  Script: {fileID: 11500000, guid: 747419968f339df40a827088f25c215d, type: 3}
 - Class: 115
   Script: {instanceID: 0}
 - Class: 120
