+++ src/ValheimVehicles/ValheimVehicles.Helpers/VehicleComponentUtils.cs
@@ -1,0 +1,6 @@
+namespace ValheimVehicles.Helpers;
+
+public class VehicleComponentUtils
+{
+  
+}
