+++ src/ValheimVehicles/ValheimVehicles.Interfaces/IRaycastPieceActivator.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+namespace ValheimVehicles.Interfaces;
+
+public interface IRaycastPieceActivator
+{
+  Transform transform { get; }
+}
