+++ src/ValheimVehicles/ValheimVehicles.Interfaces/IVehicleComponents.cs
@@ -1,0 +1,6 @@
+namespace ValheimVehicles.Interfaces;
+
+public class IVehicleComponents
+{
+  
+}
