+++ src/ValheimVehicles/ValheimVehicles.Patches/Hud_Patch.cs
@@ -59,7 +59,7 @@ public class Hud_Patch
       AnchorHud.transform.localPosition = rudderIndicator.localPosition;
   }
 
-  private static void ToggleAnchorHud(VehicleShip? vehicleShip)
+  private static void ToggleAnchorHud(VehicleBaseController? vehicleShip)
   {
     if (vehicleShip == null || vehicleShip.MovementController == null) return;
     var isAnchored = vehicleShip.MovementController.isAnchored;
@@ -94,7 +94,7 @@ public class Hud_Patch
   /// <param name="dt"></param>
   /// <param name="vehicleInterface"></param>
   public static void UpdateShipHudV2(Hud __instance, Player player, float dt,
-    VehicleShipCompat vehicleInterface)
+    VehicleControllersCompat vehicleInterface)
   {
     var speedSetting = vehicleInterface.GetSpeedSetting();
     var rudder = vehicleInterface.GetRudder();
@@ -172,7 +172,7 @@ public class Hud_Patch
     var controlledShipObj = Player_Patch.HandleGetControlledShip(player);
     if (controlledShipObj == null) return true;
 
-    var vehicleInterface = VehicleShipCompat.InitFromUnknown(controlledShipObj);
+    var vehicleInterface = VehicleControllersCompat.InitFromUnknown(controlledShipObj);
 
     if (vehicleInterface == null) return true;
