+++ src/ValheimVehicles/ValheimVehicles.Prefabs/ValheimVehicles.Prefabs.Registry/ExperimentalPrefabRegistry.cs
@@ -1,42 +1,79 @@
+using System;
 using Jotunn.Configs;
 using Jotunn.Entities;
 using Jotunn.Managers;
+using ValheimVehicles.Config;
 using ValheimVehicles.Prefabs.Registry;
 namespace ValheimVehicles.Prefabs.ValheimVehicles.Prefabs.Registry;
 
-public class CustomVehicleMastRegistry : RegisterPrefab<CustomVehicleMastRegistry>
+public class ExperimentalPrefabRegistry : RegisterPrefab<ExperimentalPrefabRegistry>
 {
-  private static void RegisterMast(string mastTier)
-  {
-    var mastAsset = LoadValheimVehicleAssets.GetMastVariant(mastTier);
-    var prefab = PrefabManager.Instance.CreateClonedPrefab(PrefabNames.GetMastByLevelName(mastTier), mastAsset);
 
-    PrefabRegistryHelpers.HoistSnapPointsToPrefab(prefab);
+  public static void OnExperimentalPrefabSettingsChange(object sender,
+    EventArgs eventArgs)
+  {
+    if (PrefabManager.Instance == null || PieceManager.Instance == null) return;
+    RegisterNautilus();
+  }
+
+  private static void RegisterNautilusVehicleShipPrefab()
+  {
+    var prefab = PrefabManager.Instance.CreateClonedPrefab(PrefabNames.Nautilus,
+      LoadValheimVehicleAssets.ShipNautilus);
     PrefabRegistryHelpers.AddNetViewWithPersistence(prefab);
-    PrefabRegistryHelpers.AddPieceForPrefab(PrefabNames.GetMastByLevelName(mastTier), prefab);
-    PrefabRegistryHelpers.SetWearNTear(prefab);
-    PieceManager.Instance.AddPiece(new CustomPiece(prefab, true,
-      new PieceConfig
-      {
-        PieceTable = PrefabRegistryController.GetPieceTableName(),
-        Category = PrefabRegistryController.SetCategoryName(VehicleHammerTableCategories.Structure),
-        Enabled = true,
-        Requirements =
-        [
-          new RequirementConfig
-          {
-            Amount = 20,
-            Item = "RoundLog",
-            Recover = true
-          }
-        ]
-      }));
+
+    var piece =
+      PrefabRegistryHelpers.AddPieceForPrefab(PrefabNames.Nautilus, prefab);
+    // var wnt = PrefabRegistryHelpers.SetWearNTear(prefab, 3);
+    // wnt.m_health = 7000f;
+    // wnt.m_burnable = false;
+    // wnt.m_supports = true;
+    // wnt.m_damages = new HitData.DamageModifiers()
+    // {
+    //   m_slash = HitData.DamageModifier.Resistant,
+    //   m_blunt = HitData.DamageModifier.Resistant,
+    //   m_pierce = HitData.DamageModifier.VeryWeak,
+    //   m_lightning = HitData.DamageModifier.VeryWeak,
+    //   m_poison = HitData.DamageModifier.Immune,
+    //   m_frost = HitData.DamageModifier.Immune,
+    //   m_fire = HitData.DamageModifier.Resistant,
+    // };
+
+    PieceManager.Instance.AddPiece(new CustomPiece(prefab, true, new PieceConfig
+    {
+      PieceTable = PrefabRegistryController.GetPieceTableName(),
+      Category = PrefabRegistryController.SetCategoryName(VehicleHammerTableCategories.Structure),
+      Enabled = true,
+      Requirements =
+      [
+        new RequirementConfig
+        {
+          Amount = 200,
+          Item = "Bronze",
+          Recover = true
+        }
+      ]
+    }));
+  }
+
+  public static void RegisterNautilus()
+  {
+    var shouldEnable =
+      PrefabConfig.AllowExperimentalPrefabs.Value;
+
+    var nautilus = PrefabManager.Instance?.GetPrefab(PrefabNames.Nautilus);
+
+    if (shouldEnable && !nautilus) RegisterNautilusVehicleShipPrefab();
+
+    if (!shouldEnable && nautilus)
+    {
+      var piece = PieceManager.Instance?.GetPiece(PrefabNames.Nautilus);
+      if (piece != null) piece.Piece.enabled = false;
+    }
   }
 
   public override void OnRegister()
   {
-    RegisterMast(PrefabNames.MastLevels.ONE);
-    RegisterMast(PrefabNames.MastLevels.TWO);
-    RegisterMast(PrefabNames.MastLevels.THREE);
+    RegisterNautilus();
   }
 }
