+++ src/ValheimVehicles/ValheimVehicles.Prefabs/ValheimVehicles.Prefabs.Registry/VehiclePrefabs.cs
@@ -182,78 +182,9 @@ public class VehiclePrefabs : IRegisterPrefab
       }));
   }
 
-  private static void RegisterNautilusVehicleShipPrefab()
-  {
-    var prefab = PrefabManager.Instance.CreateClonedPrefab(PrefabNames.Nautilus,
-      LoadValheimVehicleAssets.ShipNautilus);
-    PrefabRegistryHelpers.AddNetViewWithPersistence(prefab);
-
-    var piece =
-      PrefabRegistryHelpers.AddPieceForPrefab(PrefabNames.Nautilus, prefab);
-    // var wnt = PrefabRegistryHelpers.SetWearNTear(prefab, 3);
-    // wnt.m_health = 7000f;
-    // wnt.m_burnable = false;
-    // wnt.m_supports = true;
-    // wnt.m_damages = new HitData.DamageModifiers()
-    // {
-    //   m_slash = HitData.DamageModifier.Resistant,
-    //   m_blunt = HitData.DamageModifier.Resistant,
-    //   m_pierce = HitData.DamageModifier.VeryWeak,
-    //   m_lightning = HitData.DamageModifier.VeryWeak,
-    //   m_poison = HitData.DamageModifier.Immune,
-    //   m_frost = HitData.DamageModifier.Immune,
-    //   m_fire = HitData.DamageModifier.Resistant,
-    // };
-
-    PieceManager.Instance.AddPiece(new CustomPiece(prefab, true, new PieceConfig
-    {
-      PieceTable = PrefabRegistryController.GetPieceTableName(),
-      Category = PrefabRegistryController.SetCategoryName(VehicleHammerTableCategories.Structure),
-      Enabled = true,
-      Requirements =
-      [
-        new RequirementConfig
-        {
-          Amount = 200,
-          Item = "Bronze",
-          Recover = true
-        }
-      ]
-    }));
-  }
-
-  public void OnExperimentalPrefabSettingsChange(object sender,
-    EventArgs eventArgs)
-  {
-    if (PrefabManager.Instance == null || PieceManager.Instance == null) return;
-    RegisterNautilus();
-  }
-
-  public void RegisterNautilus()
-  {
-    var shouldEnable =
-      PrefabConfig.AllowExperimentalPrefabs.Value;
-
-    var nautilus = PrefabManager.Instance?.GetPrefab(PrefabNames.Nautilus);
-
-    if (shouldEnable && !nautilus) RegisterNautilusVehicleShipPrefab();
-
-    if (!shouldEnable && nautilus)
-    {
-      var piece = PieceManager.Instance?.GetPiece(PrefabNames.Nautilus);
-      if (piece != null) piece.Piece.enabled = false;
-    }
-  }
-
   public void Register(PrefabManager prefabManager, PieceManager pieceManager)
   {
     RegisterWaterVehicleShipPrefab();
-
-    // registers only for v3 releases.
-    // if (ValheimRaftPlugin.Version.StartsWith("3")) 
-
     RegisterLandVehiclePrefab();
-
-    RegisterNautilus();
   }
 }
